requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
Pillow==10.0.1
drf-yasg==1.21.5
python-decouple==3.8
django-rest-auth==0.9.5
django-allauth==0.54.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.22
django-countries==7.5.1
requests==2.31.0
openpyxl==3.1.2
pyyaml==6.0.1
flake8==6.0.0
black==23.7.0
isort==5.12.0
mypy==1.5.1